requirements.txt
config==0.4.0
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==1.11.15
django-bootstrap3==11.0.0
django-bootstrap4==0.0.7
django-heroku==0.3.1
django-registration==2.4.1
django-tinymce==2.7.0
gunicorn==19.9.0
Pillow==5.3.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
python-decouple==3.1
pytz==2018.6
whitenoise==4.1
